requirements.txt
#slack-cli-hooks
#pytest
#flake8
#black
slack-bolt
python-dotenv
certifi
validators
boto3